Moved out CodeStore class from AbstractCodeGenerator
This update allows us to add more function to CodeStore class.
