+++ modules/wyopcl/src/wyopcl/translator/Analyzer.java
@@ -484,7 +484,6 @@ public abstract class Analyzer {
 		return var_name;
 	}
 	
-	
-	
 
+	
 }
