+++ modules/wyopcl/src/wyopcl/translator/generator/AbstractCodeGenerator.java
@@ -208,7 +208,7 @@ public abstract class AbstractCodeGenerator {
 		// Lazy initailization.
 		if (!stores.containsKey(function)) {
 			// Put the code store into the stores
-			stores.put(function, new CodeStore(function));
+			stores.put(function, new CodeStore(function, config.isVerbose()));
 		}
 		return stores.get(function);
 	}
@@ -366,126 +366,5 @@ public abstract class AbstractCodeGenerator {
 	protected abstract void translate(IfIs code, FunctionOrMethod function);
 
 	//protected abstract void translate(SubList code, FunctionOrMethod function);
-
-	/**
-	 * Stores the generated code for a function.
-	 * 
-	 * @author Min-Hsien Weng
-	 *
-	 */
-	protected class CodeStore {
-		private String indent;
-		private FunctionOrMethod function;
-		private List<String> statements;// store the list of translated C code.
-		private HashMap<Integer, String> fields;// Stores the fields of register, e.g. 'println', 'print_s', 'println_s'
-
-		public CodeStore(FunctionOrMethod function) {
-			this.indent = "\t";
-			this.function = function;
-			this.statements = new ArrayList<String>();
-			this.fields = new HashMap<Integer, String>();
-		}
-
-		/**
-		 * Load the field to the given register.
-		 * @param reg
-		 * @param field
-		 */
-		protected void loadField(int reg, String field){
-			this.fields.put(reg, field);
-		}
-		/**
-		 * Get the field of the given register.
-		 * @param reg
-		 * @return
-		 */
-		protected String getField(int reg){
-			return this.fields.get(reg);
-		}
-
-		protected List<String> getStatements() {
-			return this.statements;
-		}
-
-		/**
-		 * Adds the statement to the list and print out the statement if the
-		 * verbose option is on.
-		 * 
-		 * @param code
-		 *            the WyIL code
-		 * @param statement
-		 *            the C code
-		 */
-		protected void addStatement(Code code, String statement) {
-			// Add the WyIL code as a comment
-			if (code != null) {
-				if (code instanceof Codes.Label) {
-					// No indentation for label bytecode
-					statements.add("//" + code.toString());
-				} else {
-					statements.add(indent + "//" + code.toString());
-				}
-
-			}
-			// Add the translated statement.
-			if (statement != null) {
-				if (config.isVerbose()) {
-					System.out.println(statement);
-				}
-				statements.add(statement);
-			}
-		}
-
-		/**
-		 * Increase the indentation.
-		 */
-		protected void increaseIndent() {
-			this.indent += "\t";
-		}
-
-		protected String getIndent() {
-			return this.indent;
-		}
-
-		/**
-		 * Decrease the indentation.
-		 */
-		protected void decreaseIndent() {
-			this.indent = this.indent.replaceFirst("\t", "");
-		}
-
-		/**
-		 * Get the type of a variable.
-		 * 
-		 * @param reg
-		 * @return
-		 */
-		protected Type getVarType(int reg) {
-			VariableDeclarations vars = function.attribute(VariableDeclarations.class);
-			Declaration declaration = vars.get(reg);
-			return declaration.type();
-		}
-
-		/**
-		 * Get the variable name of the given register
-		 * 
-		 * @param reg
-		 *            the register
-		 * @return the variable name (starting with "_")
-		 */
-		protected String getVar(int reg) {
-			VariableDeclarations vars = function.attribute(VariableDeclarations.class);
-			// Check if the register has been kept in the declarations.
-			Declaration declaration = vars.get(reg);
-			if (declaration != null) {
-				String name = declaration.name();
-				if (name != null && !name.isEmpty()) {
-					return name;
-				}
-			}
-			return prefix + reg;
-		}
-
-	}
 
 }
